+++ Accounts/DataLayer/AdoNet/PcsOiItemDao.vb
@@ -54,7 +54,7 @@ Namespace DataLayer.AdoNet
         Private Shared ReadOnly Make As Func(Of IDataReader, PcsOiItem) =
                                     Function(reader) _
             New PcsOiItem() With {
-            .AccountIdNo = Extensions.AsInt(Of Integer)(reader("AccountIdNo")),
+            .AccountIdNo = Extensions.AsInt(Of Int16)(reader("AccountIdNo")),
             .Amount = Extensions.AsDecimal(reader("Amount")),
             .ApOpenInvoiceIdNo = Extensions.AsInt(Of Integer)(reader("ApOpenInvoiceIdNo")),
             .Balance = Extensions.AsDecimal(reader("Balance")),
@@ -65,7 +65,7 @@ Namespace DataLayer.AdoNet
             .JournalCode = Extensions.AsString(reader("JournalCode")),
             .JournalIdNo = Extensions.AsId(Of Int32)(reader("JournalIdNo")),
             .PreviousBalance = Extensions.AsDecimal(reader("PreviousBalance")),
-            .Sequence = Extensions.AsInt(Of Integer)(reader("sequence")),
+            .Sequence = Extensions.AsInt(Of Int16)(reader("sequence")),
             .TransactionDate = Extensions.AsDate((reader("TransactionDate")))
             }
 
@@ -89,7 +89,7 @@ Namespace DataLayer.AdoNet
 
         Public Shared ReadOnly MakePcsOiItem As Func(Of IDataReader, PcsOiItem) = Function(reader) New PcsOiItem() With
             {
-            .AccountIdNo = Extensions.AsInt(Of Integer)(reader("AccountIdNo")),
+            .AccountIdNo = Extensions.AsInt(Of Int16)(reader("AccountIdNo")),
             .ApOpenInvoiceIdNo = Extensions.AsInt(Of Integer)(reader("IdNo")),
             .Balance = Extensions.AsDecimal(reader("Balance")),
             .IdNo = Extensions.AsInt(Of Integer)(reader("IdNo")),
